requirements.txt
asgiref==3.2.10
Django==3.1
django-appconf==1.0.4
django-autoslug==1.9.8
django-classy-tags==1.0.0
django-compressor==2.4
django-debug-toolbar==2.2
django-sekizai==1.1.0
django-taggit==1.3.0
html5lib==1.1
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
PyAIML==0.8.6
PyPDF2==1.26.0
pytz==2020.1
rcssmin==1.0.6
reportlab==3.5.48
rjsmin==1.1.0
six==1.15.0
sorl-thumbnail==12.6.3
sortable==0.1.1
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
